+++ src/utils.cpp
@@ -43,8 +43,7 @@ char *cstr(const char *str)
 
 string PathOfFile(string file)
 {
-    int i;
-
+	size_t i;
     for(i = file.length()-1; i > 0; i--)
     {
 		char c = file.at(i);
